+++ server/src/internalClusterTest/java/org/opensearch/search/aggregations/bucket/terms/StringTermsIT.java
@@ -31,7 +31,10 @@
 
 package org.opensearch.search.aggregations.bucket.terms;
 
+import com.carrotsearch.randomizedtesting.annotations.ParametersFactory;
+
 import org.opensearch.OpenSearchException;
+import org.opensearch.action.index.IndexRequestBuilder;
 import org.opensearch.action.search.SearchPhaseExecutionException;
 import org.opensearch.action.search.SearchResponse;
 import org.opensearch.common.settings.Settings;
@@ -57,13 +60,19 @@ import org.opensearch.search.aggregations.support.ValueType;
 import org.opensearch.search.builder.SearchSourceBuilder;
 import org.opensearch.test.OpenSearchIntegTestCase;
 
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Iterator;
+import java.util.List;
 import java.util.Set;
 
 import static org.opensearch.index.query.QueryBuilders.termQuery;
+import static org.opensearch.search.SearchService.CLUSTER_CONCURRENT_SEGMENT_SEARCH_SETTING;
+import static org.opensearch.search.SearchService.CONCURRENT_SEGMENT_SEARCH_PARTITION_MIN_SEGMENT_SIZE;
+import static org.opensearch.search.SearchService.CONCURRENT_SEGMENT_SEARCH_PARTITION_STRATEGY;
 import static org.opensearch.search.aggregations.AggregationBuilders.avg;
 import static org.opensearch.search.aggregations.AggregationBuilders.extendedStats;
 import static org.opensearch.search.aggregations.AggregationBuilders.filter;
@@ -83,6 +92,20 @@ public class StringTermsIT extends BaseStringTermsTestCase {
 
     public StringTermsIT(Settings staticSettings) {
         super(staticSettings);
+    }
+
+    @ParametersFactory
+    public static Collection<Object[]> parameters() {
+        return Arrays.asList(
+            new Object[] { Settings.builder().put(CLUSTER_CONCURRENT_SEGMENT_SEARCH_SETTING.getKey(), false).build() },
+            new Object[] { Settings.builder().put(CONCURRENT_SEGMENT_SEARCH_PARTITION_STRATEGY.getKey(), "segment").build() },
+            new Object[] { Settings.builder().put(CONCURRENT_SEGMENT_SEARCH_PARTITION_STRATEGY.getKey(), "force").build() },
+            new Object[] {
+                Settings.builder()
+                    .put(CONCURRENT_SEGMENT_SEARCH_PARTITION_STRATEGY.getKey(), "balanced")
+                    .put(CONCURRENT_SEGMENT_SEARCH_PARTITION_MIN_SEGMENT_SIZE.getKey(), 1000)
+                    .build() }
+        );
     }
 
     // the main purpose of this test is to make sure we're not allocating 2GB of memory per shard
@@ -1064,6 +1087,40 @@ public class StringTermsIT extends BaseStringTermsTestCase {
             f2 = terms.getBucketByKey(b2.getKeyAsString()).getAggregations().get("filter");
             assertEquals(docCount1, f1.getBuckets().get(0).getDocCount());
             assertEquals(docCount2, f2.getBuckets().get(0).getDocCount());
+        }
+    }
+
+    public void testConcurrentStringAggregation() throws Exception {
+        createIndex("test_string_terms", Settings.builder().put("index.number_of_shards", 2).put("index.number_of_replicas", 1).build());
+        try {
+            List<IndexRequestBuilder> builders = new ArrayList<>(5000);
+            for (int i = 0; i < 5; i++) {
+                builders.add(client().prepareIndex("test_string_terms").setSource("value", "val" + (i + 1)));
+            }
+            indexBulkWithSegments(builders, 2);
+            indexRandomForConcurrentSearch("test_string_terms");
+            SearchResponse response = client().prepareSearch("test_string_terms")
+                .addAggregation(
+                    terms("values").executionHint(randomExecutionHint())
+                        .field("value")
+                        .collectMode(randomFrom(SubAggCollectionMode.values()))
+                )
+                .get();
+
+            assertSearchResponse(response);
+            Terms values = response.getAggregations().get("values");
+            assertThat(values, notNullValue());
+            assertThat(values.getName(), equalTo("values"));
+            assertThat(values.getBuckets().size(), equalTo(5));
+
+            for (int i = 0; i < 5; i++) {
+                Terms.Bucket bucket = values.getBucketByKey("val" + (i + 1));
+                assertThat(bucket, notNullValue());
+                assertThat(key(bucket), equalTo("val" + (i + 1)));
+                assertThat(bucket.getDocCount(), equalTo(1L));
+            }
+        } finally {
+            internalCluster().wipeIndices("test_string_terms");
         }
     }